requirements.txt
browser-cookie3==0.19.1
certifi==2024.8.30
charset-normalizer==3.4.0
et_xmlfile==2.0.0
idna==3.10
lz4==4.3.3
numpy==2.1.1
openpyxl==3.1.5
pandas==2.2.3
pillow==10.4.0
pycryptodomex==3.21.0
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
six==1.16.0
tzdata==2024.2
urllib3==2.2.3